% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataURL.R
\name{dataURL}
\alias{dataURL}
\title{Map dataset argument to URL}
\usage{
dataURL(dataset)
}
\arguments{
\item{dataset}{Character string indicating the dataset name. Passed by \code{\link{loadSeasonalForecast}}}
}
\value{
A list of length two:
\itemize{
\item \code{URL}. URL path to the dataset
\item \code{excludeVars}. Character vector or NULL, indicating the variables in the dataset to be ignored
}
}
\description{
Map dataset name argument to the remote URL
}
\author{
J. Bedia
}
\keyword{internal}
