% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deaccumulate.R
\name{deaccumulate}
\alias{deaccumulate}
\title{Deaccumulation routine}
\usage{
deaccumulate(x, dff)
}
\arguments{
\item{x}{a vector of (accumulated) data.}

\item{dff}{Logical. Is deaccumulation performed from the first value of the time series?.}
}
\value{
a vector of deaccumulated data
}
\description{
Used for deaccumulation of precipitation in certain model data (e.g. System4).
}
\details{
When leadMonth equals 0, there is not a previous day for starting deaccumulation, and therefore the
 first value of the time series is taken 'as is'. Otherwise, one value before the start has to be taken
 to preserve time series length (this is previously done by getForecastTimeDomain.S4).
}
\author{
J. Bedia
}
\keyword{internal}
