% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deriveSurfaceSpecificHumidity.S4.R
\name{deriveSurfaceSpecificHumidity.S4}
\alias{deriveSurfaceSpecificHumidity.S4}
\title{Compute surface specific humidity in System4 datasets}
\usage{
deriveSurfaceSpecificHumidity.S4(
  gds,
  grid,
  latLon,
  runTimePars,
  memberRangeList,
  foreTimePars
)
}
\arguments{
\item{gds}{A java \dQuote{GridDataset}. This is used to load all input java \dQuote{GeoGrid}'s 
to derive the target variable.}

\item{grid}{An input java GeoGrid. This is the grid of the \sQuote{leading var} (\code{leadVar}) previously loaded
for subsetting parameter retrieval (see \code{\link{deriveInterface}} for details).}

\item{latLon}{A list of geolocation parameters, as returned by getLatLonDomainForecast}

\item{runTimePars}{A list of run time definition parameters, as returned by getRunTimeDomain.ECOMS}

\item{memberRangeList}{A list of ensemble java ranges as returned by getMemberDomain.S4}

\item{foreTimePars}{A list of forecast time definition parameters, as returned by getForecastTimeDomain.S4}
}
\value{
A n-dimensional array. Dimensions are labelled by the \dQuote{dimnames} attribute
}
\description{
Performs a basic calculation of specific humidity from sea-level pressure and
 surface temperature (for surface pressure calculation) while minimizing the amount
 of data simultaneously loaded in memory.
}
\details{
The function essentially follows the same approach as \code{\link{makeSubset.S4}}, excepting that
 at each time step it loads more than one \dQuote{GeoGrid} in order to apply the equations.
}
\references{
\url{http://meteo.unican.es/ecoms-udg/DataServer/ListOfVariables}
}
\author{
J Bedia, borrowing MatLab code from S. Herrera.
}
