% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaryLookup.ECOMS.R
\name{dictionaryLookup.ECOMS}
\alias{dictionaryLookup.ECOMS}
\title{Searches variable string in the dictionary}
\usage{
dictionaryLookup.ECOMS(dicPath, derInterface, time)
}
\arguments{
\item{dicPath}{Full path to the dictionary file (a csv file with extension \sQuote{.dic}).}

\item{derInterface}{A list as returned by \code{\link{deriveInterface}}}

\item{time}{Time specification.}
}
\value{
A data.frame of 1 row with the mapping information
}
\description{
Searches variable string provided in the dictionary to map it in the vocabulary, in order to get
all the necessary information for variable homogenization. It also includes a new column specifying the
aggregation function to be applied (if any). This funciton is ECOMS-UDG specific.
}
\details{
The function is somewhat tricky when dealing with derived variables. It applies the
time adjustment and homogenization/transformation steps indicated in the \sQuote{original variable},
(i.e., the variable to be derived) but at the end replaces the \code{shortName} by that of the
 \sQuote{leading variable} (see \code{\link{deriveInterface}} for a description of this concept),
 in order to open an existing \dQuote{GeoGrid}.
}
\references{
\url{http://meteo.unican.es/ecoms-udg/RPackage/homogeneization}
}
\author{
J. Bedia
}
