% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getForecastTimeDomain.GS5.R
\name{getForecastTimeDomain.GS5}
\alias{getForecastTimeDomain.GS5}
\title{Calculates parameters for forecast time determination along the runtime axes in GS5}
\usage{
getForecastTimeDomain.GS5(grid, dic, runTimePars, time, aggr.d, aggr.m)
}
\arguments{
\item{runTimePars}{A list of elements as returned by \code{\link{getRunTimeDomain.ECOMS}}}

\item{gcs}{a java \sQuote{GridCoordinateSystem}}
}
\value{
A list with the following elements:
\begin{itemize}
\item{forecastDates}{A list with POSIXlt dates defining the start and end of the 
representative verification time. If start and end are identical, the variable is instantaneous
and therefore the representative time interval is 0}
\item{foreTimeRangesList}{A list of length \emph{i} containing the java ranges defining the
forecast times selected along the \emph{i-th} run time axis.}
\item{deaccumFromFirst}{NULL if no deaccumulation is performed. TRUE or FALSE if deaccumulation is performed from
the first time of the runtime axis or not respectively. If FALSE, an additional runtime is added at the beginning
of each element of the runTimeList to avoid losing the first day when performing deaccumulation.}
\item{dailyAggr}{Logical. Are the forecast time values going to be used for data aggregation?. This argument is passed
to \code{makeSubset.GS5} to undertake the pertinent aggregation if TRUE.}
\end{itemize}
}
\description{
This is a subroutine of \code{\link{loadSeasonalForecast.GS5}}
}
\author{
J. Bedia
}
\keyword{internal}
