% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRunTimeDomain.CFS.R
\name{getRunTimeDomain.CFS}
\alias{getRunTimeDomain.CFS}
\title{Provides run time definition for the CFSv2 model}
\usage{
getRunTimeDomain.CFS(runDatesAll, validMonth, members, years, dataset)
}
\arguments{
\item{runDatesAll}{A POSIXlt vector with all runtime dates in the dataset}

\item{validMonth}{Integer. Initialization month.}

\item{members}{A vector defining members to select, as passed by \code{\link{loadSeasonlForecast}}}

\item{years}{Years selected (see \code{\link{loadECOMS}})}

\item{dataset}{Target hindcast dataset (see \code{\link{loadECOMS}})}
}
\value{
A list with two elements:
\itemize{
\item runDatesEnsList a list of the same length as the number of members selected, with the initialization dates
\item runTimesEnsList a list of the same length as the number of members selected, with a list of initialization
times, in the form of java ranges, as required by makeSubset.
\item years The years vector, updated if necessary. See details.
}
}
\description{
Output is passed to \code{getRunTimeDomain.ECOMS} for used-defined subsets.
}
\details{
The lagged runtime configuration of CFSv2 means that ensemble members are defined by different
 initializations, so that the dimension ensemble does not exist \emph{per se}, but is defined by
 different initializations.
 
 It may happen that there are not enough initializations for the last year to retrieve
 the data for all the initializations requested. In this case, the last year in selection
 is suppressed with a warning. If the selection corresponds to one single year, the query
 is automatically updated to the maximum number of available members, with a warning.
}
\references{
\url{http://meteo.unican.es/ecoms-udg/dataserver/datasets/CFSv2}
}
\author{
J. Bedia
}
