% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRunTimeDomain.S4.R
\name{getRunTimeDomain.S4}
\alias{getRunTimeDomain.S4}
\title{Provides run time definition for ECMWF's S4 models}
\usage{
getRunTimeDomain.S4(runDatesAll, validMonth, years)
}
\arguments{
\item{runDatesAll}{A POSIXlt vector with all runtime dates in the dataset}

\item{validMonth}{Integer. Initialization month.}

\item{years}{Numeric vector of years selected}
}
\value{
A list with two elements:
\itemize{
\item runDates a POSIXlt vector of initialization dates
\item runTimeRanges a list of of initialization times, in the form of java ranges, as required by makeSubset.
}
}
\description{
Output is passed to \code{getRunTimeDomain.ECOMS} for used-defined subsets.
}
\details{
The input/outoputs are slightly different than getRuntimeDomain.CFS because of the 
different ensemble member configuration
}
\author{
J. Bedia
}
\keyword{internal}
