% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVerticalLevelPars.ECOMS.R
\name{getVerticalLevelPars.ECOMS}
\alias{getVerticalLevelPars.ECOMS}
\title{Definition of vertical dimension slices}
\usage{
getVerticalLevelPars.ECOMS(grid, dataset, level)
}
\arguments{
\item{grid}{An object of the java class \sQuote{GeoGrid})}

\item{dataset}{dataset name}

\item{level}{Vertical level. Passed by \code{loadGridDataset}, obtained via \code{findVerticalLevel}}
}
\value{
A list with the level value and either a java Range or a java null reference
defining the index of the vertical axis (null if no vertical levels exist)
}
\description{
Returns the selected level value (if any) and a suitable java structure. This is a subroutine
of \code{loadGridDataset}, whose output is passed to \code{makeSubset}.
}
\author{
J. Bedia
}
