% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getXYcoordinatesForecast.R
\name{getXYcoordinatesForecast}
\alias{getXYcoordinatesForecast}
\title{Get geographical coordinates}
\usage{
getXYcoordinatesForecast(grid, pointXYindex, llRanges)
}
\arguments{
\item{grid}{A java \sQuote{GeoGrid}}

\item{pointXYindex}{An integer vector of length 2 with specifying the point selection (if any).}

\item{llRanges}{A list of length 2 or 1 (depending whether the subset crosses or not the dateline),
with the java \sQuote{LatLonRect} objects defining the spatial ranges.}
}
\value{
A list of the following components:
\begin{itemize}
\item{x}{x coordinates in ascending order} 
\item{y}{y coordinates in ascending order}
\item{revLat}{A logical indicating if the ordering of latitudes should be reversed or not}
\end{itemize}
}
\description{
Retrieves the geographical coordinates of the data subset, taking care of its adequate ordering
 for data retrieval and mapping
}
\author{
J Bedia
}
\keyword{internal}
