% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadGridDataset.ECOMS.R
\name{loadGridDataset.ECOMS}
\alias{loadGridDataset.ECOMS}
\title{Loads a user-defined subset of a gridded CDM dataset}
\usage{
loadGridDataset.ECOMS(var, grid, dic, level, season, years, time, latLon)
}
\description{
Loads a user-defined subset from a gridded dataset compliant with the Common
 Data Model interface
}
\references{
\url{http://www.unidata.ucar.edu/software/thredds/current/netcdf-java/tutorial/GridDatatype.html}
}
\author{
J. Bedia
}
