% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSeasonalForecast.S4.R
\name{loadSeasonalForecast.S4}
\alias{loadSeasonalForecast.S4}
\title{Load user-defined subsets from System4 datasets}
\usage{
loadSeasonalForecast.S4(
  dataset,
  gds,
  var,
  grid,
  dic,
  members,
  latLon,
  runTimePars,
  time,
  level,
  aggr.d,
  aggr.m,
  derInterface
)
}
\description{
Load user-defined subsets from System4 datasets considering the ensemble dimensions and
other particular characteristics of System4 datasets. This is a subroutine of
 \code{\link{loadSeasonalForecast}}.
 
 @param dataset character string indicating the dataset requested.
 @param var character string indicating the variable to download.
 @param grid A java GeoGrid.
 @param dic A single-row data.frame, as returned by \code{dictionaryLookup.ECOMS}.
 @param members Numeric vector indicating the members to be retrieved.
 @param latLon A list of geolocation parameters as returned by \code{getLatLonDomainForecast}
 @param runTimePars A list of parameters defining de initializations to be taken and other.
 auxiliary parameters, as returned by \code{getRunTimeDomain.ECOMS}.
 @param time Verification time. See \code{\link{loadSeasonalForecast}}
 @param derInterface A list of components indicating the interface for derived variables,
  when relevant. See details.
 @return A list of components, as returned by \code{\link{loadSeasonalForecast}}.
 @details In the case of derived variables, these are computed on-the-fly applying the
  appropriate method. The idea is computing the derived variable loading the minimum 
  object sizes at a time in order to optimize the available memory. Thus, different
  variants of the \code{\link{makeSubset.S4}} method (referred to as \sQuote{interfaces} in the dictionary
  and henceforth within the code) have been implemented, and called alternatively depending on the 
  variable to be derived. They operate in a time-slice basis, removing from the memory the input variables
  at each time step within the for loop once the derived variable is calculated,
  freeing as much space as possible. The different functions for deriving variables are named with the
  \sQuote{derive} prefix.
 @references \url{http://meteo.unican.es/ecoms-udg/ListOfVariables}
 @author J. Bedia 
 @importFrom loadeR C4R.vocabulary
 @importFrom loadeR dictionaryTransformForecast
}
