% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSubset.CFS.R
\name{makeSubset.CFS}
\alias{makeSubset.CFS}
\title{Makes a logical subset of a CFSv2 GeoGrid}
\usage{
makeSubset.CFS(grid, latLon, runTimePars, foreTimePars)
}
\arguments{
\item{grid}{An input java GeoGrid}

\item{latLon}{A list of geolocation parameters, as returned by getLatLonDomainForecast}

\item{runTimePars}{A list of run time definition parameters, as returned by \code{getRunTimeDomain.ECOMS}}

\item{foreTimePars}{A list of forecast time definition parameters, as returned by \code{getForecastTimeDomain.CFS}}
}
\value{
A list with the n-dimensional array of data and the modified foreTimePars with adjusted dates depending on the
temporal aggregation performed.
}
\description{
Makes a logical subset of a CFSv2 GeoGrid using the parameters specified by the user,
applying the java methods makeSubset and readDataSlice. Subroutine of \code{loadSeasonalForecast.CFS}
}
\details{
Dimensions of length one are dropped and the \dQuote{dimnames} attribute is consequently modified.
In the current version the Z dimension is ignored (and dropped), as it is not planned to include multi-level variables
in the ECOMS-UDG by the moment. Because of the lagged-runtime configuration of CFSv2 for member definition,
the dimension \sQuote{ensemble} doen not exist. In turn, this is created and included as a dimension in the returned array
from the run time parameters passed by runTimePars and foreTimePars.

This function performs the temporal aggregations.
}
\references{
\url{http://www.unidata.ucar.edu/software/thredds/v4.3/netcdf-java/v4.3/javadocAll/ucar/nc2/dt/grid/GeoGrid.html}
}
\author{
J Bedia and A. Cofi\~no
}
