\name{world_lines}

\alias{world_lines}

\docType{data}

\title{

Map of the world land areas

}

\description{

A \link[sp]{SpatialLines} object representing the land areas of the world.

}

\usage{data(world_lines)}



\details{

This dataset is useful as a spatial reference for data visualization.

}

\source{

Made with Natural Earth. Free vector and raster map data \@ naturalearthdata.com
\url{www.naturalearthdata.com/http//www.naturalearthdata.com/download/110m/physical/ne_110m_land.zip}

}


\keyword{datasets}
